# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Voltpilot-EMS Portal API
  version: 0.1.0-mvp
  description: >
    OpenAPI contract for the portal backend (Spring Boot). BINDING contract
    (architecture section 20 item 8). Covers the MVP portal surface: auth-guarded
    tenancy, device claiming, telemetry and schedule views, and economic KPIs
    (architecture section 4).


    Every non-system endpoint requires a Keycloak-issued Bearer JWT (realm
    `voltpilot`). The `tenant_id` claim scopes all data via Postgres
    Row-Level-Security - a caller only ever sees its own tenant's rows; another
    tenant's resources return 404 (not found) rather than 403. The
    sites/devices/telemetry/prices/weather/schedule and device-claim endpoints
    are implemented in services/api; KPIs remain a stub pending a later
    increment.
  contact:
    name: Voltpilot-EMS
servers:
  - url: http://localhost:8090
    description: Local dev (docker compose / `mvn spring-boot:run`)
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/sites:
    get:
      tags: [sites]
      summary: List sites for the caller's tenant (tenant_id from OIDC token).
      responses:
        "200":
          description: Sites (only the caller's tenant, via RLS).
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Site" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [sites]
      summary: Create a site for the caller's own tenant (self-service onboarding).
      description: >
        Creates a site (Standort) for the CALLER's tenant. The tenant is taken from
        the OIDC `tenant_id` claim (via TenantContext) and enforced by RLS' WITH
        CHECK - never from the request body - so a customer can only ever create a
        site for themselves. This unblocks onboarding: a fresh customer creates a
        site, then claims devices into it. Implemented in services/api.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSiteRequest" }
      responses:
        "201":
          description: Site created (for the caller's tenant).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Site" }
        "400": { description: Invalid input (blank name, out-of-range coordinates, or bad zone). }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/sites/{siteId}:
    put:
      tags: [sites]
      summary: Update a site (name/bidding zone/coordinates).
      description: >
        Validation mirrors create. RLS makes a foreign site a 404 and pins the
        row to its tenant, so a site can never be moved across tenants. Admins
        edit any tenant's sites through the same endpoint via the X-Tenant-Id
        tenant switcher.
      parameters:
        - { name: siteId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSiteRequest" }
      responses:
        "200":
          description: Updated site.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Site" }
        "400": { description: Invalid input (blank name, out-of-range coordinates, or bad zone). }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Site not found for this tenant. }
    delete:
      tags: [sites]
      summary: Delete a site (guarded; cascades assets + series data).
      description: >
        Refused (409) while the site still has devices - they must be removed
        first. On success the site's assets cascade by FK and its series data
        (telemetry, rollups, forecast, schedule, weather, forecast-quality rows)
        is deleted in the same transaction. Use the deletion-preview endpoint to
        show the concrete consequences before confirming.
      parameters:
        - { name: siteId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Site and its recorded data deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Site not found for this tenant. }
        "409": { description: The site still has devices; remove them first. }
  /api/v1/sites/{siteId}/deletion-preview:
    get:
      tags: [sites]
      summary: What deleting this site would remove (for the confirm dialog).
      parameters:
        - { name: siteId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Device count and recorded-data extent/counts.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SiteDeletionPreview" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Site not found for this tenant. }
...
  /api/v1/sites/{siteId}/telemetry:
    get:
      tags: [telemetry]
      summary: Query telemetry timeseries for a site.
      parameters:
        - { name: siteId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: from, in: query, required: false, schema: { type: string, format: date-time } }
        - { name: to, in: query, required: false, schema: { type: string, format: date-time } }
      responses:
        "200":
          description: Telemetry points (tenant-scoped via RLS). `from`/`to` default to the last 24h.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TelemetryPoint" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Site not found for this tenant. }
  /api/v1/sites/{siteId}/prices:
    get:
      tags: [market]
      summary: Day-ahead spot prices (15-min slots) for a site's bidding zone.
      description: >
        Returns the day-ahead electricity spot prices for the site's bidding zone
        (DE-LU). Prices are market-wide PER ZONE (not per tenant), but the endpoint
        is site-scoped so the portal reads "the price for this site": the site is
        resolved via RLS (a foreign site is 404), then its zone's series is
        returned. Collected KEYLESS from energy-charts.info (Fraunhofer ISE), with
        ENTSO-E as an alternative source. `from`/`to` default to roughly
        today+tomorrow. Implemented in services/api.
      parameters:
        - { name: siteId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: from, in: query, required: false, schema: { type: string, format: date-time } }
        - { name: to, in: query, required: false, schema: { type: string, format: date-time } }
      responses:
        "200":
          description: Price series for the site's zone.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PriceSeries" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Site not found for this tenant. }
  /api/v1/sites/{siteId}/weather:
    get:
      tags: [weather]
      summary: Latest weather forecast (hourly) for a site.
      description: >
        Returns the most recent weather-forecast run for the site (temperature,
        cloud cover and shortwave/GHI radiation useful for PV), hourly over the
        coming days. The `weather_forecast` table carries `tenant_id` and is
        RLS-scoped, and the site lookup is RLS-gated, so a foreign site is 404 and
        no other tenant's weather can leak. Collected KEYLESS from Open-Meteo
        (EU-hosted). Implemented in services/api.
      parameters:
        - { name: siteId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Latest weather forecast (empty points array if none stored yet).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WeatherForecast" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Site not found for this tenant. }
  /api/v1/sites/{siteId}/schedule:
    get:
      tags: [schedules]
      summary: Current 24h optimization schedule (15-min slots) for a site.
      description: >
        The latest optimizer run from the schedule hypertable (written by
        services/optimization): per-slot planned battery power (+charge /
        -discharge), projected grid power and SoC, the day-ahead price and the
        projected cost vs. the no-battery baseline, plus the headline savingsEur
        over the horizon. RLS-scoped by tenant. Empty slots array (null
        metadata) when no plan has been produced yet.
      parameters:
        - { name: siteId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The site's current plan (empty slots array if none yet).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Site not found for this tenant. }
  /api/v1/sites/{siteId}/history:
    get:
      tags: [history]
      summary: A site's history ("Historie") for one period, money lens first.
      description: >
        Bucketed pv/load/grid/SoC series (15-min for day, hourly for week,
        Europe/Berlin-daily for month/year; served from the telemetry rollup
        tables, day live from raw telemetry) with the matching day-ahead price
        and per-bucket import cost, plus period totals: consumption / PV /
        import / export kWh, grid cost EUR (import energy x matching 15-min
        price), battery savings EUR (persisted optimizer plans, latest run per
        slot, null where none exist), Autarkiegrad and Eigenverbrauchsquote.
        For range=day the response additionally carries the Tagesprotokoll
        (notable events in plain German) and the persisted plan trajectory for
        the plan-vs-actual overlay. Period boundaries are Europe/Berlin local
        time. RLS-scoped by tenant.
      parameters:
        - { name: siteId, in: path, required: true, schema: { type: string, format: uuid } }
        - name: range
          in: query
          required: false
          schema: { type: string, enum: [day, week, month, year], default: day }
        - name: at
          in: query
          required: false
          schema: { type: string, format: date }
          description: Any date inside the wanted period (default today, Europe/Berlin).
      responses:
        "200":
          description: The site's history for the period (empty buckets if no data).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/History" }
        "400": { description: Unknown range value. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Site not found for this tenant. }
  /api/v1/sites/{siteId}/forecast-quality:
    get:
      tags: [forecast-quality]
      summary: Forecast-model quality ("Prognosequalität") for a site.
      description: >
        The shadow-mode forecasting transparency view (docs/forecasting.md):
        which forecast model is LIVE per kind (activeLoadModel/activePvModel,
        resolved from server config - the same env the optimizer reads), the
        lifecycle of every model that has run on the site (challengers report
        "collecting: day X of N" while their self-gate is unmet; trained models
        carry trainedAt/trainRows and the top feature importances with
        plain-German labels), the per-model daily error series (MAE kW, nMAE %,
        bias, skill vs the baseline - positive = better) computed by the daily
        evaluation job, and the daily plan-vs-actual economics. Challenger
        promotion is NEVER automatic; it is an operator env flip informed by
        this data. All rows come from RLS-scoped tables - a foreign site is a
        404; a fresh site yields empty (but well-formed) lists.
      parameters:
        - { name: siteId, in: path, required: true, schema: { type: string, format: uuid } }
        - name: days
          in: query
          required: false
          schema: { type: integer, default: 30, minimum: 1, maximum: 90 }
          description: How many trailing days of accuracy/plan series to return.
      responses:
        "200":
          description: The site's forecast-quality view.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForecastQuality" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Site not found for this tenant. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Keycloak-issued OIDC access token (realm `voltpilot`). Must carry a
        `tenant_id` claim; the API sets it as the Postgres `app.tenant_id`
        session variable to drive Row-Level-Security.

        Tenant switcher (Portal-Admins only): a token with the `platform-admin`
        realm role (which carries NO tenant_id) may select a tenant per request
        via the `X-Tenant-Id` header (UUID). The customer endpoints then serve
        that tenant's data through the same RLS scoping the customer gets. The
        header is IGNORED for every non-admin token, so it can never widen a
        customer's access; an admin without the header sees zero rows
        (default-deny).
  responses:
    Unauthorized:
      description: Missing, expired or invalid Bearer token.
...
  schemas:
...
    Site:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        biddingZone: { type: string, example: DE-LU }
        latitude:
          type: number
          nullable: true
          description: WGS84 latitude; ties the site to its weather forecast.
          example: 52.52
        longitude:
          type: number
          nullable: true
          description: WGS84 longitude.
          example: 13.405
...
    CreateSiteRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, description: Site display name (required, non-blank). }
        biddingZone:
          type: string
          enum: [DE-LU, AT, CH]
          description: "Day-ahead bidding zone; defaults to 'DE-LU'."
        latitude:
          type: number
          nullable: true
          minimum: -90
          maximum: 90
          description: Optional WGS84 latitude; needed for the site's weather forecast.
        longitude:
          type: number
          nullable: true
          minimum: -180
          maximum: 180
          description: Optional WGS84 longitude.
    UpdateSiteRequest:
      type: object
      required: [name]
      description: Same fields and validation as CreateSiteRequest.
      properties:
        name: { type: string, description: Site display name (required, non-blank). }
        biddingZone:
          type: string
          enum: [DE-LU, AT, CH]
          description: "Day-ahead bidding zone; defaults to 'DE-LU'."
        latitude: { type: number, nullable: true, minimum: -90, maximum: 90 }
        longitude: { type: number, nullable: true, minimum: -180, maximum: 180 }
    SiteDeletionPreview:
      type: object
      description: What deleting a site would remove (drives the confirm dialog).
      properties:
        deviceCount: { type: integer, description: Devices still at the site; > 0 blocks the delete (409). }
        telemetryCount: { type: integer, format: int64 }
        telemetryFrom: { type: string, format: date-time, nullable: true }
        telemetryTo: { type: string, format: date-time, nullable: true }
        forecastCount: { type: integer, format: int64 }
        scheduleCount: { type: integer, format: int64 }
        weatherCount: { type: integer, format: int64 }
...
    TelemetryPoint:
      type: object
      properties:
        ts: { type: string, format: date-time }
        powerKw: { type: number }
        socPct: { type: number }
        pvPowerKw: { type: number }
        loadKw: { type: number }
        gridLimitKw:
          type: number
          description: Observed effective §14a power limit.
    PriceSeries:
      type: object
      description: Day-ahead spot prices for a bidding zone (market-wide, not per tenant).
      properties:
        biddingZone: { type: string, example: DE-LU }
        resolution:
          type: string
          nullable: true
          description: ISO-8601 slot resolution of the latest data.
          example: PT15M
        currency: { type: string, example: EUR }
        points:
          type: array
          items: { $ref: "#/components/schemas/PricePoint" }
    PricePoint:
      type: object
      properties:
        ts: { type: string, format: date-time, description: Slot start (UTC). }
        end: { type: string, format: date-time, description: Slot end (exclusive). }
        priceEurMwh: { type: number, description: Spot price in EUR/MWh. }
    WeatherForecast:
      type: object
      properties:
        runAt:
          type: string
          format: date-time
          nullable: true
          description: Issue time of the forecast run (null if none stored yet).
        points:
          type: array
          items: { $ref: "#/components/schemas/WeatherPoint" }
    WeatherPoint:
      type: object
      properties:
        ts: { type: string, format: date-time, description: Forecast target hour (UTC). }
        temperatureC: { type: number, nullable: true, description: 2 m air temperature. }
        cloudCoverPct: { type: number, nullable: true, description: Total cloud cover 0..100. }
        ghiWM2:
          type: number
          nullable: true
          description: Shortwave / global-horizontal irradiance (W/m²), PV-relevant.
        dniWM2: { type: number, nullable: true, description: Direct (beam) radiation (W/m²). }
        dhiWM2: { type: number, nullable: true, description: Diffuse radiation (W/m²). }
    Schedule:
      type: object
      properties:
        planId: { type: string, format: uuid, nullable: true }
        deviceId: { type: string, format: uuid, nullable: true, description: "Executing battery device (the MQTT schedule topic's device)." }
        generatedAt: { type: string, format: date-time, nullable: true }
        slotMinutes: { type: integer, example: 15 }
        savingsEur:
          type: number
          nullable: true
          description: Projected savings over the horizon vs. leaving the battery idle (the no-battery baseline).
        slots:
          type: array
          items:
            type: object
            properties:
              start: { type: string, format: date-time }
              batteryKw: { type: number, description: "Planned battery power; + = charge, - = discharge." }
              gridKw: { type: number, description: "Projected net grid power; + = import, - = export." }
              socPct: { type: number, description: "Planned state of charge at slot end." }
              priceEurMwh: { type: number }
              costEur: { type: number, description: "Projected slot cost with the plan." }
              baselineCostEur: { type: number, description: "Projected slot cost with the battery idle." }
...